refactor: Move main file directly to feature
