+++ lib/core/style/progress_theme_extensions.dart
@@ -2,11 +2,9 @@ import 'package:flutter/material.dart';
 
 class ProgressThemeExtension extends ThemeExtension<ProgressThemeExtension> {
   ProgressThemeExtension({
-    required this.levelBarGradient,
     required this.streakStarGradient,
   });
 
-  final Gradient? levelBarGradient;
   final Gradient? streakStarGradient;
 
   @override
@@ -15,7 +13,6 @@ class ProgressThemeExtension extends ThemeExtension<ProgressThemeExtension> {
     Gradient? streakStarGradient,
   }) =>
       ProgressThemeExtension(
-        levelBarGradient: levelBarGradient,
         streakStarGradient: streakStarGradient,
       );
 
@@ -28,11 +25,6 @@ class ProgressThemeExtension extends ThemeExtension<ProgressThemeExtension> {
       return this;
     }
     return ProgressThemeExtension(
-      levelBarGradient: Gradient.lerp(
-        levelBarGradient,
-        other.levelBarGradient,
-        t,
-      ),
       streakStarGradient: Gradient.lerp(
         streakStarGradient,
         other.streakStarGradient,
